+++ docs/changelog.md
@@ -2,7 +2,15 @@
 
 ## Changelog
 
+### v0.4.0 Rename the script and change the user interface
+
+- Change the script's file extension from `.pyw` to `.py` and implement a command line-only UI.
+
+Based on PyWriter 2.14.4
+
+
 ### v0.2.1 Service release
+
 * Change the Markdown linefeeds.
 * Add a customization template